Fix "tainted canvas" security issue for mask loading
If the mask image was delivered over CORS, which happens in cases
of an S3 assetstore or webpack dev server environment, we could
not read data out of the canvas after the image had been loaded
into it for security reasons. It wasn't sufficient to set the
crossOrigin attribute on the fetched image, because we actually
do need to use cookie authentication to download them as they
are private data stored in Girder.